# omod/src/main/webapp/resources/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API for Messages Module.",
    "version" : "v1",
    "title" : "Messages Module REST API"
  },
  "basePath" : "/openmrs/ws",
...
  "paths" : {
...
    "/messages/countryProperty/setValues" : {
      "post" : {
        "tags" : [ "REST API for Country Property Actions" ],
        "summary" : "None",
        "description" : "Set the values for Country Property. Creates properties if needed.",
        "operationId" : "setValue",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CountryPropertyValueList"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Country Property Values have been set.",
            "schema" : {
              "$ref" : "#/definitions/PatientTemplatesWrapper"
            }
          },
          "500" : {
            "description" : "Failed to set Country Property value."
          }
        }
      }
    },
...
    "/messages/triggerMessage/{personUuid}/{templateNames}/{channelType}" : {
      "post" : {
        "tags" : [ "REST API for message trigger actions" ],
        "summary" : "None",
        "description" : "Triggers sending of an instant message. Creates required message entities.",
        "operationId" : "triggerSendMessage",
        "parameters" : [ {
          "name" : "personUuid",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "templateNames",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "channelType",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Message has been triggered"
          },
          "500" : {
            "description" : "Failed to trigger message"
          }
        }
      }
    }
  },
  "definitions" : {
...
    "CountryPropertyValueDTO" : {
      "type" : "object",
      "properties" : {
        "country" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        }
      }
    },
    "CountryPropertyValueList" : {
      "type" : "object",
      "properties" : {
        "values" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CountryPropertyValueDTO"
          }
        }
      }
    },
...
    "PatientTemplateDTO" : {
      "type" : "object",
      "required" : [ "actorId", "patientId", "templateId" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "templateFieldValues" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TemplateFieldValueDTO"
          }
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "templateId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "actorId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "actorTypeId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "uuid" : {
          "type" : "string"
        }
      }
    },
    "PatientTemplatesWrapper" : {
      "type" : "object",
      "properties" : {
        "patientTemplates" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PatientTemplateDTO"
          }
        }
      }
    },
...
    "TemplateFieldValueDTO" : {
      "type" : "object",
      "required" : [ "templateFieldId" ],
      "properties" : {
        "uuid" : {
          "type" : "string"
        },
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "templateFieldId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "value" : {
          "type" : "string"
        }
      }
    },
...
  }
}